+++ packages/web3-common/src/constants.ts
@@ -45,4 +45,7 @@ export const ERR_INVALID_PROVIDER = 601;
 export const ERR_INVALID_CLIENT = 602;
 export const ERR_SUBSCRIPTION = 603;
 
+export const ERR_PRIVATE_KEY_LENGTH = 701;
+export const ERR_INVALID_PRIVATE_KEY = 702;
+
 export const GENESIS_BLOCK_NUMBER = '0x0';