+++ packages/web3-eth-abi/src/api/parameters_api.ts
@@ -8,7 +8,7 @@ import { formatParam, isAbiFragment, mapTypes, modifyParams } from '../utils';
 /**
  * Should be used to encode list of params
  */
-export const encodeParameters = (abi: AbiInput[], params: unknown[]): string => {
+export const encodeParameters = (abi: ReadonlyArray<AbiInput>, params: unknown[]): string => {
 	try {
 		const modifiedTypes = mapTypes(Array.isArray(abi) ? abi : [abi]);
 		const modifiedParams: Array<unknown> = [];